Incorporate review updates.
